fixed abstract class, added projectile offset
