+++ Assets/Scripts/Player/BaseTestClass.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8d649720e2a13f040b61304d1b8dd728
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
