+++ Assets/Scripts/Player/BaseTestClass/BaseTestClass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1471a5774a07375459ea3b6d99779ca5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
