+++ Assets/Scripts/Player/PlayerController.cs
@@ -2,11 +2,9 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-
-public class PlayerController : CharacterController
+public abstract class PlayerController : CharacterController
 {
     protected Rigidbody2D rb;
-    [SerializeField] protected SpriteRenderer spriteRenderer;
 
     protected override void Start()
     {
@@ -14,28 +12,17 @@ public class PlayerController : CharacterController
         currentHealth = maxHealth;
     }
 
-    void FixedUpdate()
+    protected override void Update()
     {
-        base.FixedUpdate();
+        base.Update();
     }
 
     protected override void Move()
     {
-        float horizontalInput = Input.GetAxisRaw("Horizontal");
-        float verticalInput = Input.GetAxisRaw("Vertical");
+        float horizontalInput = Input.GetAxis("Horizontal");
+        float verticalInput = Input.GetAxis("Vertical");
         Vector2 movement = new Vector2(horizontalInput, verticalInput).normalized;
         rb.velocity = movement * moveSpeed;
-
-        if (horizontalInput < 0f)
-        {
-            spriteRenderer.flipX = true;
-
-        }
-        else if (horizontalInput > 0f)
-        {
-            spriteRenderer.flipX = false;
-        }
     }
-
 
 }
